+++ src/components/Profile/BaanSelection/components/InviteLink.tsx
@@ -1,0 +1,30 @@
+import { useAuth } from '@/context/AuthContext';
+import { Square2StackIcon } from '@heroicons/react/24/outline';
+
+export default function InviteLink() {
+    const { group } = useAuth();
+
+    const inviteLink = `${
+        process.env.NEXT_PUBLIC_APP_BASE_URL
+    }/?join=${encodeURIComponent(
+        group ? group.token : 'GROUP_TOKEN_NOT_FOUND'
+    )}`;
+
+    function handleLinkCopy() {
+        navigator.clipboard.writeText(`${inviteLink}`);
+    }
+    return (
+        <div className="flex w-full flex-row items-center justify-center gap-4 text-white">
+            <label className="min-w-max text-xl lg:text-2xl">Invite Link</label>
+            <div className="relative flex w-full">
+                <div className="w-full rounded-full bg-white px-5 py-2 text-sm text-purple ring-8 ring-white/30">
+                    {inviteLink}
+                </div>
+                <Square2StackIcon
+                    className="absolute right-2 top-1/2 w-7 -translate-y-1/2 cursor-pointer rounded-full bg-pink-400 p-1"
+                    onClick={handleLinkCopy}
+                />
+            </div>
+        </div>
+    );
+}
